added sku and refactoring

+++ src/main/java/com/spring/mongocrud/controllers/ProductController.java
@@ -3,14 +3,19 @@ package com.spring.mongocrud.controllers;
 import com.spring.mongocrud.exceptions.AlreadyPresentException;
 import com.spring.mongocrud.models.Product;
 import com.spring.mongocrud.services.ProductService;
+import com.spring.mongocrud.validator.ProductRequestResponse;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
+import javax.validation.Valid;
+import java.util.*;
 
 @RestController
 @AllArgsConstructor
@@ -20,45 +25,76 @@ public class ProductController {
     @Autowired private ProductService productService;
 
     @GetMapping("products/getAllProducts")
-    public List<Product> getAllProducts() {
+    public ResponseEntity<?> getAllProducts() {
         List<Product> productList = productService.getAllProducts();
-        return productList;
+
+
+        return ResponseEntity.ok(productList);
     }
 
     @GetMapping("/products")
-    public List<Product> getActiveProducts() {
+    public ResponseEntity<?> getActiveProducts() {
         List<Product> productList = productService.getActiveProducts();
         //System.out.println("productList = " + productList);
-        return productList;
+        ArrayList<ProductRequestResponse> productRequestResponseArrayList = new ArrayList<>();
+        productList.forEach(product -> {
+            if (!product.getStatus().toString().equalsIgnoreCase("deleted"))
+                productRequestResponseArrayList.add(new ProductRequestResponse(product));
+        });
+        return ResponseEntity.ok(productList);
     }
 
     @PostMapping("/products/add")
-    public ResponseEntity<?> addProduct(@RequestBody Product product) {
+    public ResponseEntity<?> addProduct(@Valid @RequestBody ProductRequestResponse productRequestResponse) {
+        Product newProduct;
         try {
-            productService.addProduct(product);
+            newProduct = productService.addProduct(productRequestResponse);
         } catch (AlreadyPresentException ex) {
-            return ResponseEntity.ok("Id already exist in the database...!");
+            return ResponseEntity.internalServerError().body("Id already exist in the database...!");
         }
-        return ResponseEntity.ok("Product Added");
+        return ResponseEntity.of(Optional.of(newProduct));
     }
 
     @PutMapping("/products/update")
-    public ResponseEntity<?> updateProduct(@RequestBody Product product) {
-        productService.updateProduct(product);
-        return ResponseEntity.ok("Product Updated");
+    public ResponseEntity<?> updateProduct(@Valid @RequestBody ProductRequestResponse productRequestResponse) {
+        Product updatedProduct = productService.updateProduct(productRequestResponse);
+        productRequestResponse = new ProductRequestResponse(updatedProduct);
+        return ResponseEntity.of(Optional.of(productRequestResponse));
     }
 
     @DeleteMapping("products/delete/{id}")
     public ResponseEntity<?> deleteProduct(@PathVariable String id) {
-        productService.deleteProduct(id);
-        return ResponseEntity.ok("Product with id:" + id + " Deleted Successfully");
+        if (id != null) {
+            Product product = productService.deleteProduct(id);
+            ProductRequestResponse productRequestResponse = new ProductRequestResponse(product);
+            return ResponseEntity.of(Optional.of(productRequestResponse));
+        } else {
+            return ResponseEntity.ok(null);
+        }
     }
 
     @PutMapping("products/softDelete/{id}")
     public ResponseEntity<?> softDelete(@PathVariable String id) {
-        productService.softDelete(id);
-        return ResponseEntity.ok("Soft Deleted product with id " + id);
+        if (id != null) {
+            Product product = productService.softDelete(id);
+            ProductRequestResponse productRequestResponse = new ProductRequestResponse(product);
+            return ResponseEntity.ok(productRequestResponse);
+        } else {
+            return ResponseEntity.ok(null);
+        }
     }
 
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public Map<String, String> handleValidationExceptions(MethodArgumentNotValidException ex) {
+        Map<String, String> errors = new HashMap<>();
+        ex.getBindingResult().getAllErrors().forEach((error) -> {
+            String fieldName = ((FieldError) error).getField();
+            String errorMessage = error.getDefaultMessage();
+            errors.put(fieldName, errorMessage);
+            errors.put("Status", "Bad Request(400)");
+        });
+        return errors;
+    }
 
 }
